Move to using int for regions
I'll have to rewrite RegionMapping when I do merging, but for now its fine as is.

+++ modules/core/src/main/java/at/redi2go/photonics/core/rendering/world/block/BlockModel.java
@@ -12,6 +12,6 @@ public interface BlockModel extends Disposable {
         Vector3i offset();
         int boundingVolume();
 
-        BlockEntry toEntry(short region);
+        BlockEntry toEntry(int region);
     }
 }
